Partly migrate `write_conference_create_request` to `rasn`.
This actually required that a bug be fixed with rasn's PER generation
first. Once this was done, output matched the original function
except for a single octet. I believe this octet was a byte count
for the following data, and that in the original code it was
one less than it was meant to be.